+++ app/src/androidTest/java/com/team7/cmput301/android/theirisproject/RegisterActivityTest2.java
@@ -1,126 +1,88 @@
 package com.team7.cmput301.android.theirisproject;
 
-import android.app.Activity;
-import android.support.test.espresso.Espresso;
-import android.support.test.filters.LargeTest;
-import android.support.test.rule.ActivityTestRule;
-import android.support.test.runner.AndroidJUnit4;
+import android.test.ActivityInstrumentationTestCase2;
+import android.widget.Button;
+import android.widget.EditText;
 
+import com.robotium.solo.Solo;
+import com.team7.cmput301.android.theirisproject.R;
 import com.team7.cmput301.android.theirisproject.activity.RegisterActivity;
 
-import org.junit.Before;
-import org.junit.Rule;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-
-import static android.support.test.espresso.Espresso.onView;
-import static android.support.test.espresso.action.ViewActions.*;
-import static android.support.test.espresso.assertion.ViewAssertions.matches;
-import static android.support.test.espresso.matcher.RootMatchers.*;
-import static android.support.test.espresso.matcher.ViewMatchers.*;
-import static org.hamcrest.Matchers.is;
-import static org.hamcrest.Matchers.not;
-
 /**
- * RegisterActivityTest contains tests pertaining to registering users in RegisterActivity
+ * Contains register tests for RegisterActivity
+ *
  *
  * @author Jmmxp
  * @see RegisterActivity
  */
+public class RegisterActivityTest2 extends ActivityInstrumentationTestCase2<RegisterActivity> {
 
-@RunWith(AndroidJUnit4.class)
-@LargeTest
-public class RegisterActivityTest2 {
+    // TODO: The tests in here do not work due to Solo ClassNotFound exceptions and the like
+    // We have to decide if we want to use Robotium or Espresso
 
-    private Activity activity;
+    private Solo solo;
 
-    @Rule
-    public ActivityTestRule<RegisterActivity> activityRule = new ActivityTestRule<>(RegisterActivity.class);
-
-    @Before
-    public void setUp() {
-        activity = activityRule.getActivity();
+    public RegisterActivityTest2() {
+        super(RegisterActivity.class);
     }
 
-    // Test registration when not all fields have been filled out
-    @Test
-    public void testRegisterIncomplete() {
-        onView(withId(R.id.register_button))
-                .perform(click());
-
-        checkToast(getString(R.string.register_incomplete));
-
-        fillEditText(R.id.name_edit_text, "John Doe");
-
-
-        closeKeyboardAndWait();
-        onView(withId(R.id.register_button))
-                .perform(click());
-
-        checkToast(getString(R.string.register_incomplete));
-    }
-
-    // Currently only works if the account has already been registered
-
-    /**
-     * Test registration when all fields are filled out but the email is already in use
-     */
-    @Test
-    public void testRegisterAlreadyRegistered() {
-        fillEditText(R.id.name_edit_text, "John Doe");
-        fillEditText(R.id.password_edit_text, "badpassword");
-        fillEditText(R.id.email_edit_text, "johndoe@gmail.com");
-        fillEditText(R.id.phone_edit_text, "123-456-7890");
-
-        closeKeyboardAndWait();
-        onView(withId(R.id.register_button))
-                .perform(click());
-
-        checkToast(getString(R.string.register_failure));
-    }
-
-    /**
-     * Test registration when all fields are filled out and user is registered successfully
-     */
-    @Test
-    public void testRegisterSuccess() {
-        fillEditText(R.id.name_edit_text, "John Doe");
-        fillEditText(R.id.password_edit_text, "badpassword");
-        fillEditText(R.id.email_edit_text, "johndoe2@gmail.com");
-        fillEditText(R.id.phone_edit_text, "123-456-7890");
-
-        closeKeyboardAndWait();
-        onView(withId(R.id.register_button))
-                .perform(click());
-
-        // TODO: Delete test user from the DB, or use a mock DB before doing this test
-        // checkToast(getString(R.string.register_success));
-    }
-
-    private void fillEditText(int editTextResId, String text) {
-        onView(withId(editTextResId))
-                .perform(typeText(text));
-    }
-
-    private void closeKeyboardAndWait() {
-        Espresso.closeSoftKeyboard();
-        try {
-            Thread.sleep(250);
-        } catch (InterruptedException e) {
-            e.printStackTrace();
-        }
-    }
-
-    // Check if a Toast displays given text
-    // REFERENCE: https://stackoverflow.com/questions/28390574/checking-toast-message-in-android-espresso
-    private void checkToast(String toastText) {
-        onView(withText(toastText))
-                .inRoot(withDecorView(not(is(activity.getWindow().getDecorView()))))
-                .check(matches(isDisplayed()));
-    }
-
-    private String getString(int stringResId) {
-        return activity.getString(stringResId);
-    }
+//    @Override
+//    protected void setUp() throws Exception {
+//        super.setUp();
+//        solo = new Solo(getInstrumentation(), getActivity());
+//    }
+//
+//    @Override
+//    protected void tearDown() throws Exception {
+//        super.tearDown();
+//        solo.finishOpenedActivities();
+//    }
+//
+//    public void testActivity() {
+//        String activityName = RegisterActivity.class.getSimpleName();
+//        solo.assertCurrentActivity("Wrong activity found, should be " + activityName, activityName);
+//    }
+//
+//    // Test registration when not all fields have been filled out
+//    public void testRegisterIncomplete() {
+//        Button registerButton = (Button) solo.getView(R.id.register_button);
+//
+//        solo.clickOnView(registerButton);
+//
+//        assertTrue(solo.waitForText(getString(R.string.register_failure)));
+//
+//        // Try registering with only the name field filled out
+//        EditText nameEditText = (EditText) solo.getView(R.id.name_edit_text);
+//        solo.enterText(nameEditText, "John Doe");
+//
+//        solo.clickOnView(registerButton);
+//
+//        assertTrue(solo.waitForText(getString(R.string.register_success)));
+//    }
+//
+    // Test registration when all fields have been filled out
+//    public void testRegisterComplete() {
+//        Button registerButton = (Button) solo.getView(R.id.register_button);
+//
+//
+//        // Try registering with only the name field filled out
+//        EditText nameEditText = (EditText) solo.getView(R.id.name_edit_text);
+//        EditText passwordEditText = (EditText) solo.getView(R.id.password_edit_text);
+//        EditText emailEditText = (EditText) solo.getView(R.id.email_edit_text);
+//        EditText phoneEditText = (EditText) solo.getView(R.id.phone_edit_text);
+//
+//        solo.enterText(nameEditText, "John Doe");
+//        solo.enterText(passwordEditText, "badpassword");
+//        solo.enterText(emailEditText, "johndoe@gmail.com");
+//        solo.enterText(phoneEditText, "123-456-7890");
+//
+//        solo.clickOnView(registerButton);
+//
+//        assertTrue(solo.waitForText(getString(R.string.register_failure)) || solo.waitForText(getString(R.string.register_incomplete)));
+//    }
+//
+//    private String getString(int stringId) {
+//        return getActivity().getString(stringId);
+//    }
 
 }
